Display Stocks added to StockMain page
